Migrated App.jsx to typeScript
