+++ src/pages/NotFound.tsx
@@ -1,4 +1,9 @@
 import React from "react";
-const NotFound = () => <h1>Not found</h1>;
+const NotFound = () => (
+  <>
+    {" "}
+    <h1>Not found</h1>
+  </>
+);
 
 export default NotFound;
